+++ src/physics/backend/rapier.rs
@@ -13,21 +13,13 @@
 //!     .add_plugins(TiledPhysicsPlugin::<TiledPhysicsRapierBackend>::default());
 //! ```
 
+use crate::prelude::*;
 use bevy::prelude::*;
-use bevy_ecs_tilemap::{anchor::TilemapAnchor, map::TilemapGridSize};
 use bevy_rapier2d::{
     prelude::*,
     rapier::prelude::{Isometry, Real, SharedShape},
 };
 use tiled::{ObjectLayerData, ObjectShape};
-
-use crate::{
-    names::TiledNameFilter,
-    physics::collider::{ColliderCreated, TiledCollider},
-    tiled::{event::TiledEvent, helpers::grid_size_from_map, map::asset::TiledMapAsset},
-};
-
-use super::{TiledPhysicsBackend, TiledPhysicsBackendOutput};
 
 /// The [`TiledPhysicsBackend`] to use for Rapier 2D integration.
 #[derive(Default, Reflect, Copy, Clone, Debug)]