+++ CHANGELOG.md
@@ -17,6 +17,11 @@ A [migration guide](https://adrien-bon.github.io/bevy_ecs_tiled/migrations/v0_8.
 - Add helpers functions to the `TiledMapStorage` component so it's easier to work with.
 - Several files have been moved around or renamed.
 
+### Documentation
+
+- Overall improvement of the documentation
+- Add some example to enable `physics` and `user_properties` features (#100)
+
 ## v0.7.4
 
 ### Features